+++ src/posts/posts.service.ts
@@ -166,6 +166,8 @@ export class PostsService {
   }
   async deletePost(id: string) {
     const postId = new Types.ObjectId(id);
+    const post = await this.postsRepository.findPostById(postId);
+    if(!post) throw new HttpException('', HttpStatus.NOT_FOUND)
     if(!postId) throw new HttpException('', HttpStatus.NOT_FOUND)
     return await this.postsRepository.delete(postId);
   }
